refactor(#23): Move typescript files to seperate classes
