+++ typescript/main.ts
@@ -1,0 +1,2 @@
+export const DEST_PATH = './docs/Schedules/';
+export const ROUTES_THAT_NEED_UPDATES_DIR = './docs/routes/';
